# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Alumni Tracking System API",
    "version": "1.0.0",
    "description": "Web Programlama dersi 3. Hafta API dokümantasyonu"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Sunucu durumunu kontrol eder",
        "responses": {
          "200": {
            "description": "Başarılı"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Tüm kullanıcıları listeler",
        "responses": {
          "200": {
            "description": "Başarılı"
          }
        }
      },
      "post": {
        "summary": "Yeni bir kullanıcı ekler",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "graduationYear": { "type": "number" },
                  "company": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kullanıcı Eklendi"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "ID'ye göre tek kullanıcı getirir",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı"
          }
        }
      },
      "put": {
        "summary": "Kullanıcıyı günceller",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "company": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı"
          }
        }
      },
      "delete": {
        "summary": "Kullanıcıyı siler",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Silindi"
          }
        }
      }
    }
  }
}